+++ src/main/java/com/promptoven/reviewBatchService/application/AggregateService.java
@@ -1,0 +1,8 @@
+package com.promptoven.reviewBatchService.application;
+
+import com.promptoven.reviewBatchService.dto.out.AggregateResponseDto;
+
+public interface AggregateService {
+
+    AggregateResponseDto getAggregateData(String productUuid);
+}
